requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-groq==0.1.5
langchain-mongodb==0.1.5
langgraph==0.0.64
python-dotenv==1.0.0
pymongo==4.6.1
pydantic==2.6.3
aiohttp==3.9.3

# Testing
pytest==7.4.4
